Trait pallet_receipts::Config

pub trait Config: Config {
    type DomainHash: Parameter + Member + MaybeSerializeDeserialize + Debug + MaybeDisplay + SimpleBitOps + Ord + Default + Copy + CheckEqual + Hash + AsRef<[u8]> + AsMut<[u8]> + MaxEncodedLen;
    type RuntimeEvent: From<Event<Self>> + IsType<<Self as Config>::RuntimeEvent>;
    type MaximumReceiptDrift: Get<Self::BlockNumber>;
    type ReceiptsPruningDepth: Get<Self::BlockNumber>;
}

Configuration trait of this pallet.

Implement this type for a runtime in order to customize this pallet.

Required Associated Types

type DomainHash: Parameter + Member + MaybeSerializeDeserialize + Debug + MaybeDisplay + SimpleBitOps + Ord + Default + Copy + CheckEqual + Hash + AsRef<[u8]> + AsMut<[u8]> + MaxEncodedLen

Domain block hash type.

type RuntimeEvent: From<Event<Self>> + IsType<<Self as Config>::RuntimeEvent>

type MaximumReceiptDrift: Get<Self::BlockNumber>

Maximum execution receipt drift.

If the primary number of an execution receipt plus the maximum drift is bigger than the best execution chain number, this receipt will be rejected as being too far in the future.

type ReceiptsPruningDepth: Get<Self::BlockNumber>

Number of execution receipts kept in the state.